+++ tests/06_html_with_comments/src/script1.js
@@ -1,0 +1,6 @@
+{
+	const $p = document.createElement('p');
+	$p.innerHTML = 'Output from <span class="highlight">script1</span>.';
+	document.getElementById('content').appendChild($p);
+}
+