+++ tests/06_html_with_comments/src/script2.js
@@ -1,0 +1,6 @@
+{
+	const $p = document.createElement('p');
+	$p.innerHTML = 'Output from <span class="highlight">script2</span>. This script should be ignored.';
+	document.getElementById('content').appendChild($p);
+}
+